TODO-10 : guarding against deleting todos that don't exist

+++ webapp.py
@@ -34,6 +34,8 @@ def update(todo_id):
 @app.route("/delete/<int:todo_id>")
 def delete(todo_id):
     todo = Todo.query.filter_by(id=todo_id).first()
+    if not todo:
+        return "Todo not found", 404
     db.session.delete(todo)
     db.session.commit()
     return redirect(url_for("home"))
